+++ packages/riverpod/lib/riverpod.dart
@@ -24,6 +24,7 @@ export 'src/framework.dart'
         FamilyCreate,
         AsyncSelector,
         FamilyBase,
+        FamilyOverrideImpl,
         AutoDisposeProviderElementMixin,
         FamilyOverride,
         NotifierFamilyBase,