+++ packages/riverpod/lib/src/state_provider/auto_dispose.dart
@@ -39,6 +39,20 @@ class AutoDisposeStateProvider<T> extends _StateProviderBase<T> {
   )
   @override
   late final Refreshable<StateController<T>> state = _state(this);
+
+  /// {@macro riverpod.overridewith}
+  Override overrideWith(
+    Create<T, AutoDisposeStateProviderRef<T>> create,
+  ) {
+    return ProviderOverride(
+      origin: this,
+      override: AutoDisposeStateProvider<T>(
+        create,
+        from: from,
+        argument: argument,
+      ),
+    );
+  }
 }
 
 /// The element of [StateProvider].
@@ -59,4 +73,18 @@ class AutoDisposeStateProviderFamily<R, Arg> extends AutoDisposeFamilyBase<
     super.name,
     super.dependencies,
   }) : super(providerFactory: AutoDisposeStateProvider.new);
+
+  /// {@macro riverpod.overridewith}
+  Override overrideWith(
+    R Function(AutoDisposeStateProviderRef<R> ref, Arg arg) create,
+  ) {
+    return FamilyOverrideImpl<R, Arg, AutoDisposeStateProvider<R>>(
+      this,
+      (arg) => AutoDisposeStateProvider<R>(
+        (ref) => create(ref, arg),
+        from: from,
+        argument: arg,
+      ),
+    );
+  }
 }