+++ packages/riverpod/lib/src/stream_provider/auto_dispose.dart
@@ -36,6 +36,20 @@ class AutoDisposeStreamProvider<T> extends _StreamProviderBase<T>
 
   @override
   late final Refreshable<Stream<T>> stream = _stream(this);
+
+  /// {@macro riverpod.overridewith}
+  Override overrideWith(
+    Create<Stream<T>, AutoDisposeStreamProviderRef<T>> create,
+  ) {
+    return ProviderOverride(
+      origin: this,
+      override: AutoDisposeStreamProvider<T>(
+        create,
+        from: from,
+        argument: argument,
+      ),
+    );
+  }
 }
 
 /// The element of [AutoDisposeStreamProvider].
@@ -61,4 +75,18 @@ class AutoDisposeStreamProviderFamily<R, Arg> extends AutoDisposeFamilyBase<
     super.name,
     super.dependencies,
   }) : super(providerFactory: AutoDisposeStreamProvider.new);
+
+  /// {@macro riverpod.overridewith}
+  Override overrideWith(
+    Stream<R> Function(AutoDisposeStreamProviderRef<R> ref, Arg arg) create,
+  ) {
+    return FamilyOverrideImpl<AsyncValue<R>, Arg, AutoDisposeStreamProvider<R>>(
+      this,
+      (arg) => AutoDisposeStreamProvider<R>(
+        (ref) => create(ref, arg),
+        from: from,
+        argument: arg,
+      ),
+    );
+  }
 }