+++ packages/riverpod/test/providers/async_notifier/async_notifier_test.dart
@@ -805,6 +805,51 @@ void main() {
     });
   }
 
+  test('supports overrideWithNotifier', () {
+    final provider = AsyncNotifierProvider<AsyncTestNotifier<int>, int>(
+      () => AsyncTestNotifier((ref) => 0),
+    );
+    final autoDispose = AsyncNotifierProvider.autoDispose<
+        AutoDisposeAsyncTestNotifier<int>, int>(
+      () => AutoDisposeAsyncTestNotifier((ref) => 0),
+    );
+    final container = createContainer(
+      overrides: [
+        provider.overrideWithNotifier(() => AsyncTestNotifier((ref) => 42)),
+        autoDispose.overrideWithNotifier(
+          () => AutoDisposeAsyncTestNotifier((ref) => 84),
+        ),
+      ],
+    );
+
+    expect(container.read(provider).value, 42);
+    expect(container.read(autoDispose).value, 84);
+  });
+
+  test('supports family overrideWith', () {
+    final family =
+        AsyncNotifierProvider.family<AsyncTestNotifierFamily<int>, int, int>(
+      () => AsyncTestNotifierFamily<int>((ref) => 0),
+    );
+    final autoDisposeFamily = AsyncNotifierProvider.autoDispose
+        .family<AutoDisposeAsyncTestNotifierFamily<int>, int, int>(
+      () => AutoDisposeAsyncTestNotifierFamily<int>((ref) => 0),
+    );
+    final container = createContainer(
+      overrides: [
+        family.overrideWithNotifier(
+          () => AsyncTestNotifierFamily<int>((ref) => 42),
+        ),
+        autoDisposeFamily.overrideWithNotifier(
+          () => AutoDisposeAsyncTestNotifierFamily<int>((ref) => 84),
+        ),
+      ],
+    );
+
+    expect(container.read(family(10)).value, 42);
+    expect(container.read(autoDisposeFamily(10)).value, 84);
+  });
+
   group('AutoDispose variant', () {
     test('can watch autoDispose providers', () {
       final dep = Provider.autoDispose((ref) => 0);