+++ packages/riverpod/test/providers/future_provider/future_provider_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_types_on_closure_parameters
+
 import 'dart:async';
 
 import 'package:mockito/mockito.dart';
@@ -7,6 +9,42 @@ import 'package:test/test.dart';
 import '../../utils.dart';
 
 void main() {
+  test('supports overrideWith', () {
+    final provider = FutureProvider<int>((ref) => 0);
+    final autoDispose = FutureProvider.autoDispose<int>((ref) => 0);
+    final container = createContainer(
+      overrides: [
+        provider.overrideWith((FutureProviderRef<int> ref) => 42),
+        autoDispose.overrideWith(
+          (AutoDisposeFutureProviderRef<int> ref) => 84,
+        ),
+      ],
+    );
+
+    expect(container.read(provider).value, 42);
+    expect(container.read(autoDispose).value, 84);
+  });
+
+  test('supports family overrideWith', () {
+    final family = FutureProvider.family<String, int>((ref, arg) => '0 $arg');
+    final autoDisposeFamily = FutureProvider.autoDispose.family<String, int>(
+      (ref, arg) => '0 $arg',
+    );
+    final container = createContainer(
+      overrides: [
+        family.overrideWith(
+          (FutureProviderRef<String> ref, int arg) => '42 $arg',
+        ),
+        autoDisposeFamily.overrideWith(
+          (AutoDisposeFutureProviderRef<String> ref, int arg) => '84 $arg',
+        ),
+      ],
+    );
+
+    expect(container.read(family(10)).value, '42 10');
+    expect(container.read(autoDisposeFamily(10)).value, '84 10');
+  });
+
   test('Emits AsyncLoading before the create function is executed', () async {
     final container = createContainer();
     late AsyncValue<int> state;
@@ -329,6 +367,7 @@ void main() {
       final container = createContainer(
         parent: root,
         overrides: [
+          // ignore: deprecated_member_use_from_same_package
           provider.overrideWithProvider(FutureProvider((ref) async => 42)),
         ],
       );