+++ packages/riverpod/test/providers/notifier/notifier_test.dart
@@ -350,6 +350,50 @@ void main() {
     }
   }
 
+  test('supports overrideWithNotifier', () {
+    final provider = NotifierProvider<TestNotifier<int>, int>(
+      () => TestNotifier((ref) => 0),
+    );
+    final autoDispose =
+        NotifierProvider.autoDispose<AutoDisposeTestNotifier<int>, int>(
+      () => AutoDisposeTestNotifier((ref) => 0),
+    );
+    final container = createContainer(
+      overrides: [
+        provider.overrideWithNotifier(() => TestNotifier((ref) => 42)),
+        autoDispose.overrideWithNotifier(
+          () => AutoDisposeTestNotifier((ref) => 84),
+        ),
+      ],
+    );
+
+    expect(container.read(provider), 42);
+    expect(container.read(autoDispose), 84);
+  });
+
+  test('supports family overrideWith', () {
+    final family = NotifierProvider.family<TestNotifierFamily<int>, int, int>(
+      () => TestNotifierFamily<int>((ref) => 0),
+    );
+    final autoDisposeFamily = NotifierProvider.autoDispose
+        .family<AutoDisposeTestNotifierFamily<int>, int, int>(
+      () => AutoDisposeTestNotifierFamily<int>((ref) => 0),
+    );
+    final container = createContainer(
+      overrides: [
+        family.overrideWithNotifier(
+          () => TestNotifierFamily<int>((ref) => 42),
+        ),
+        autoDisposeFamily.overrideWithNotifier(
+          () => AutoDisposeTestNotifierFamily<int>((ref) => 84),
+        ),
+      ],
+    );
+
+    expect(container.read(family(10)), 42);
+    expect(container.read(autoDisposeFamily(10)), 84);
+  });
+
   group('modifiers', () {
     void canBeAssignedToAlwaysAliveRefreshable<T>(
       AlwaysAliveRefreshable<T> provider,