+++ packages/riverpod/test/providers/state_provider/state_provider_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_types_on_closure_parameters
+
 import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
@@ -5,6 +7,44 @@ import 'package:test/test.dart';
 import '../../utils.dart';
 
 void main() {
+  test('supports overrideWith', () {
+    final provider = StateProvider<int>(
+      (ref) => 0,
+    );
+    final autoDispose = StateProvider.autoDispose<int>(
+      (ref) => 0,
+    );
+    final container = createContainer(
+      overrides: [
+        provider.overrideWith((StateProviderRef<int> ref) => 42),
+        autoDispose.overrideWith((AutoDisposeStateProviderRef<int> ref) => 84),
+      ],
+    );
+
+    expect(container.read(provider), 42);
+    expect(container.read(autoDispose), 84);
+  });
+
+  test('supports family overrideWith', () {
+    final family = StateProvider.family<String, int>((ref, arg) => '0 $arg');
+    final autoDisposeFamily = StateProvider.autoDispose.family<String, int>(
+      (ref, arg) => '0 $arg',
+    );
+    final container = createContainer(
+      overrides: [
+        family.overrideWith(
+          (StateProviderRef<String> ref, int arg) => '42 $arg',
+        ),
+        autoDisposeFamily.overrideWith(
+          (AutoDisposeStateProviderRef<String> ref, int arg) => '84 $arg',
+        ),
+      ],
+    );
+
+    expect(container.read(family(10)), '42 10');
+    expect(container.read(autoDisposeFamily(10)), '84 10');
+  });
+
   test(
       'on refresh, does not notify listeners if the new value is identical to the previous one',
       () {
@@ -250,6 +290,7 @@ void main() {
       final container = createContainer(
         parent: root,
         overrides: [
+          // ignore: deprecated_member_use_from_same_package
           provider.overrideWithProvider(
             StateProvider((ref) => 42),
           ),